Moved the script so that it would always load
